requirements.txt
boto3==1.35.66
dask==2024.8.2
glob2==0.7
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.2
tqdm==4.66.5
Unidecode==1.3.8